+++ tests/testthat/test-compounds.R
@@ -174,9 +174,8 @@ test_that("reporting works", {
     for (grp in names(compoundTable(comps)))
         checkmate::expect_file_exists(getWorkPath("compounds", sprintf("%s-%s.pdf", class(fGroups), grp)))
     
-    expect_file(reportMD(fGroups, getWorkPath(), reportPlots = "none",
-                         compounds = comps, MSPeakLists = plists),
-                getWorkPath("report.html"))
+    expect_reportMD(makeReportMD(fGroups, reportPlots = "none",
+                                 compounds = comps, MSPeakLists = plists))
 })
 
 test_that("reporting empty objects works", {
@@ -184,9 +183,8 @@ test_that("reporting empty objects works", {
     expect_error(reportCSV(fGroups, getWorkPath(), compounds = compsEmpty), NA)
     expect_error(reportPDF(fGroups, getWorkPath(), reportFGroups = FALSE, compounds = compsEmpty,
                            MSPeakLists = plistsEmpty), NA)
-    expect_file(reportMD(fGroups, getWorkPath(), reportPlots = "none",
-                         compounds = compsEmpty, MSPeakLists = plistsEmpty),
-                getWorkPath("report.html"))
+    expect_reportMD(makeReportMD(fGroups, reportPlots = "none",
+                                 compounds = compsEmpty, MSPeakLists = plistsEmpty))
 })
 
 test_that("plotting works", {